Use px instead of % to define borer-radius
Each corner uses two dimentions to render border-radius. Width and height of the element.

When using % and the element is not a perfect square, you end up with a weird proportion in the radius renderization.

To learn more about it:
https://developer.mozilla.org/en-US/docs/Web/CSS/border-radius#Values

When using px to define border-radius, browser will use same value to both corner edges resulting in a smooth round corner.